Renames and moves tests. Alters npm test to run all tests
